requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
anyio==4.8.0
argcomplete==3.6.2
attrs==25.4.0
audioop-lts==0.2.2
beautifulsoup4==4.14.2
blivet==3.12.1
blivet-gui==2.6.0
boto3==1.40.54
botocore==1.40.54
Brlapi==0.8.6
cairocffi==1.7.1
CairoSVG==2.7.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.1
crypt_r==3.13.1
cssselect2==0.8.0
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
dbus_next==0.2.3
defusedxml==0.7.1
discord.py==2.6.4
distro==1.9.0
dnf==4.23.0
file-magic==0.4.0
frozenlist==1.8.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
humanize==4.12.0
idna==3.10
initial-setup==0.3.101
iso639==0.1.4
jmespath==1.0.1
langtable==0.0.69
libcomps==0.1.22
libdnf==0.74.0
libdnf5==5.2.16.0
louis==3.33.0
lxml==5.3.2
mpmath==1.3.0
multidict==6.7.0
nftables==0.1
olefile==0.47
packaging==24.2
pexpect==4.9.0
pid==2.2.3
pillow==11.1.0
ply==3.11
praw==7.8.1
prawcore==2.4.0
productmd==1.45
propcache==0.4.1
psutil==6.1.1
ptyprocess==0.7.0
pwquality==1.4.5
PyAudio==0.2.13
pycairo==1.25.1
pycparser==2.20
pycups==2.0.4
pyenchant==3.2.2
PyGObject==3.50.0
pykickstart==3.62
pyparted==3.13.0
PySocks==1.7.1
pystray==0.17.3
python-augeas==1.2.0
python-dateutil==2.8.2
python-dotenv==1.1.1
python-gettext==4.0
python-manatools==0.0.4
python-meh==0.52
python-xlib==0.33
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.2
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.4
requests-file==2.0.0
requests-ftp==0.3.1
rpm==4.20.1
s3transfer==0.14.0
selinux @ file:///builddir/build/BUILD/libselinux-3.8-build/libselinux-3.8/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.8-build/selinux-3.8/python/sepolicy
setools==4.5.1
setroubleshoot @ file:///builddir/build/BUILD/setroubleshoot-3.3.35-build/setroubleshoot-3.3.35/src
setuptools==74.1.3
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
sos==4.10.0
soupsieve==2.7
sympy==1.14.0
systemd-python==235
tinycss2==1.4.0
typing_extensions==4.12.2
update-checker==0.18.0
urllib3==2.3.0
webencodings==0.5.1
websocket-client==1.9.0
wikipedia==1.4.0
xkbregistry==0.3
yarl==1.22.0
